Feat: differentiate the fiat / non-fiat collaterals for the USDC / wETH pools of Maple-v2
